Add files via upload
This is a Bash script that automates the installation of various tools for Android penetration testing. The script installs the following tools:
1- Frida
2- Drozer
3- APKTool
4- ADB
5- MobSF

The script checks for the dependencies required by each tool, and installs them if they are missing. It also provides the option to install individual tools or all of them at once. The script is useful for anyone who wants to set up a complete Android penetration testing environment quickly and easily.
To use this script, simply download or clone the repository from GitHub and run the "android-pentesting-installer.sh" file. The script will prompt the user to select which tools they would like to install and will guide them through the installation process.

This script is useful for security researchers, penetration testers, and anyone interested in Android security testing. It saves time and effort by automating the installation process of popular Android security tools. It can be used as a standalone script or integrated into a larger workflow.